Fix field error diagnostics being mapped to 'endOfInput' in Parser monad. Fixes #99.
Until now, cassava parsed CSV lines analogous to this:

     !hdr <- header (decDelimiter opts)
     vals <- records hdr
     _ <- optional endOfLine
     endOfInput
     let !v = V.fromList vals
     return (hdr, v)

with

    records :: Parser [a] -- fake signature inserted by me
    records = do
      r <- record
      case convertToUserType r of
        Left msg -> fail msg
        Right val ->
          vals <- (endOfLine *> records hdr) <|> pure []
          return (val : vals)

The problem here is that the `fail msg` would always be ignored by
attoparsec, unless it happened in the first line.

This is because if there are multiple lines, they are parsed using
`(endOfLine *> records hdr) <|> pure []`.
If the `(endOfLine *> records hdr)` part fails (e.g. with `fail msg`),
then attoparsec will throw away its error message and try with
`pure []` instead (which will fail later with `endOfInput`).

The solution is to first acknowledge that `records :: Parser [a]` can
only know that it's done (and attach the eventual `[]`) when it
observes that the file is over.
Consequently, the setup

    vals <- records hdr
    endOfInput

does not make sense.

Second, by switching from

    parseOtherLines <|> endOfInput

to

    endOfInput <|> parseOtherLines

we can ensure that the (potential) error of the entire expression
is the one of `parseOtherLines`.

In general, it seems always advisable to check for `endOfInput`
first, before trying other parser alternatives.